Added remove book

+++ insides/book.py
@@ -1,9 +1,9 @@
-# Aprasoma knyga
+# book.py
 class Book:
-    def __init__(self, pavadinimas: str, autorius: str, isleidimo_metai: float, zanras: str) -> None:
+    def __init__(self, pavadinimas: str, autorius: str, isleidimo_metai: int, zanras: str) -> None:
         self.pavadinimas = pavadinimas
         self.autorius = autorius
-        self.isleidimo_metai = isleidimo_metai
+        self.isleidimo_metai = isleidimo_metai  # Use the correct attribute name
         self.zanras = zanras
 
     def __str__(self):
